tool_integration_service/requirements.txt
fastapi
uvicorn
pydantic[email]
sqlalchemy
python-dotenv
mysql-connector-python
httpx
pytest
coverage
passlib
